Fix the bug in weather
When user use the apps for the first time there is no current city
saved on the NSUserDefault. Since we are not using the location
services, I have hard coded the value of “Melbourne” to automatically
load the weather from Melbourne. Not a better solution but it will do
for a last minute fix